# modules/web/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "TokenResponse": {
        "description": "OAuth2-compatible access token payload for login, refresh, OAuth, and SSO.\n\nLocal login sets ``refresh_token`` to ``None``. Supabase paths populate both\naccess and refresh tokens when Auth issues a session.\n\nAttributes:\n    access_token: Bearer JWT for protected API routes.\n    token_type: Always ``bearer`` for clients that expect OAuth2 shape.\n    expires_in: Access-token lifetime in seconds (``>= 1``).\n    refresh_token: Optional Supabase refresh token for rotation/logout.",
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "minimum": 1.0,
            "title": "Expires In",
            "type": "integer"
          },
          "refresh_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Token"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "expires_in"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      },
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Save Ma Money API",
    "version": "0.0.0-contract"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "description": "OAuth2 password flow — Supabase Auth sign-in or local credential verify.\n\nClients must put the **email** in the OAuth2 ``username`` form field when\n``AUTH_PROVIDER=supabase``. Local mode still accepts email or derived username.\nSupabase login provisionally ensures a Papita row by Auth ``sub``; recent\norphan Auth users may be deleted if provision fails (not for inactive tenants).\n\nArgs:\n    form: OAuth2 password form (``username`` + ``password``).\n    _rate_limit: Per-IP login rate-limit dependency (side effect only).\n    settings: Selects Auth mode and JWT/session defaults.\n    users_service: Verifies credentials (local) or ensures Auth-linked user.\n    auth_manager: Issues HS256 access tokens in local mode only.\n\nReturns:\n    ``TokenResponse`` with access token; Supabase also returns ``refresh_token``.\n\nRaises:\n    HTTPException: 401 for bad credentials; 429 when Auth rate-limits; other\n        provision/Auth failures mapped from helpers.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_api_v1_auth_login_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "Revoke the session: Supabase Auth sign-out and/or Redis JWT denylist.\n\n* ``AUTH_PROVIDER=supabase`` — calls Supabase ``sign_out``, then denylists the\n  access token when Redis is enabled so remaining JWT TTL cannot be reused.\n* ``AUTH_PROVIDER=local`` — when Redis is enabled, denylists the access token\n  and returns 204; otherwise returns HTTP 501 (no refresh/session store).\n\nAccess token may be supplied in the JSON body or as ``Authorization: Bearer``.\nPer-IP OAuth/auth rate limits apply (logout may run before tenant context).\n\nArgs:\n    body: Refresh token (required for Supabase) and optional access token.\n    settings: Application settings selecting Auth provider.\n    credentials: Optional bearer credentials for the access JWT.\n    session_store: Redis-backed JWT denylist (no-op when Redis disabled).\n    _rate_limit: Per-IP auth rate-limit dependency (side effect only).\n\nReturns:\n    Empty 204 response on success, or deferred JSON with status 501 in local\n    mode without Redis.\n\nRaises:\n    HTTPException: 400 when access token missing; 401 on Auth errors;\n        429 when the per-IP auth limit is exceeded; 503 when Auth misconfigured.",
        "operationId": "logout_api_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Logout",
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Rotate Supabase Auth access/refresh tokens.\n\nWhen ``AUTH_PROVIDER=local``, returns HTTP 501 (no refresh-token store).\n\nArgs:\n    body: Current Supabase refresh token.\n    _rate_limit: Per-IP OAuth/refresh rate-limit dependency (side effect only).\n    settings: Application settings selecting Auth provider.\n\nReturns:\n    New access and refresh tokens from Supabase Auth, or a deferred JSON body\n    with status 501 in local mode.\n\nRaises:\n    HTTPException: 401 on invalid refresh token; 400 on malformed Auth response;\n        503 when Supabase URL/anon key are missing.",
        "operationId": "refresh_token_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Token",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "description": "Register a new user via Supabase Auth (or local ``UsersService``).\n\nEmail is the canonical login identity. A schema ``username`` handle is\nderived from the email local-part when not supplied. Under Supabase, Auth\nsignup runs first; Papita provision is idempotent. Provision failure after a\nsuccessful Auth create triggers best-effort Admin orphan cleanup.\n\nArgs:\n    body: Registration payload (email, password, optional profile fields).\n    _rate_limit: Per-IP register rate-limit dependency (side effect only).\n    settings: Selects ``supabase`` vs ``local`` Auth mode.\n    users_service: Creates or links the tenant ``users`` row.\n\nReturns:\n    Public ``RegisterResponse`` for the new or already-linked tenant user,\n    including ``email_confirmation_required`` when Confirm email is on.\n\nRaises:\n    HTTPException: Mapped Auth/provision errors (409 conflict, 401 inactive,\n        429 rate limit, 503 misconfigured Supabase, etc.).",
        "operationId": "register_user_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register User",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/resend-confirmation": {
      "post": {
        "description": "Resend Supabase signup confirmation email (Bearer / smoke twin of BFF).\n\nAlways returns 204 on soft success (including unknown emails). Rate-limited\nwith the register bucket because both trigger Auth SMTP.",
        "operationId": "resend_confirmation_api_v1_auth_resend_confirmation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendConfirmationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resend Confirmation",
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/api/v1/bff/auth/login": {
      "post": {
        "description": "Authenticate and set an HttpOnly session-id cookie (JWTs stay server-side).",
        "operationId": "bff_login_api_v1_bff_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BffLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BffSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bff Login",
        "tags": [
          "BFF Authentication"
        ]
      }
    },
...
    "/api/v1/bff/auth/register": {
      "post": {
        "description": "Register via the same Auth/UsersService path as ``POST /auth/register``.\n\nDoes not create a BFF session — clients should call ``POST /bff/auth/login``\nafter email confirmation when required (PPT-068). Never sets ``papita_sid``.",
        "operationId": "bff_register_api_v1_bff_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BffRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bff Register",
        "tags": [
          "BFF Authentication"
        ]
      }
    },
...
    "/api/v1/bff/auth/session": {
      "get": {
        "description": "Probe BFF session for SPA bootstrap.\n\nReturns ``authenticated: false`` (200) when no valid cookie/Bearer — avoids\ntreating anonymous bootstrap as an error. When authenticated via cookie,\nincludes ``csrf_token`` for mutation headers.",
        "operationId": "bff_session_api_v1_bff_auth_session_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BffSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Bff Session",
        "tags": [
          "BFF Authentication"
        ]
      }
    },
...
  }
}